LayoutTests/imported/w3c/web-platform-tests/xhr/send-authentication-basic-setrequestheader-existing-session-expected.txt


FAIL XMLHttpRequest: send() - "Basic" authenticated request using setRequestHeader() when there is an existing session assert_equals: expected "08095d5b-696e-47c7-beb5-154f56219747\npass" but got "adab1597-e232-41f4-9fc6-cac3b4bbf046\nopen-pass"
Note: this test will only work as expected once per browsing session. Restart browser to re-test.
